--- a/.rspec
+++ b/.rspec
@@ -1,1 +1,1 @@
---require rails_helper
+--require spec_helper