--- a/spec/i18n_spec.rb
+++ b/spec/i18n_spec.rb
@@ -15,12 +15,10 @@ end
 RSpec.describe 'I18n' do
   let(:i18n) { I18n::Tasks::BaseTask.new }
   let(:unused_keys) { i18n.unused_keys }
+  let(:unused_key_count) { unused_keys.leaves.count }
 
   let(:failure_msg) do
-    <<-MSG.squish
-      #{unused_keys.leaves.count} unused i18n keys, run `i18n-tasks unused' to
-      show them
-    MSG
+    "#{unused_key_count} unused i18n keys, run `i18n-tasks unused' to show them"
   end
 
   it 'does not have unused keys' do