--- a/spec/requests/stylesheets_spec.rb
+++ b/spec/requests/stylesheets_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe "Stylesheets", type: :request do
 
   require "sprockets"