--- a/spec/unit/abstract_view_factory_spec.rb
+++ b/spec/unit/abstract_view_factory_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 require 'active_admin/abstract_view_factory'
 
 RSpec.describe ActiveAdmin::AbstractViewFactory do