--- a/spec/unit/action_builder_spec.rb
+++ b/spec/unit/action_builder_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe 'defining actions from registration blocks', type: :controller do
   let(:klass){ Admin::PostsController }