--- a/spec/unit/active_admin_spec.rb
+++ b/spec/unit/active_admin_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin do
   %w(register register_page unload! load! routes).each do |method|
     it "delegates ##{method} to application" do