--- a/spec/unit/application_spec.rb
+++ b/spec/unit/application_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require 'fileutils'
 
 RSpec.describe ActiveAdmin::Application do