--- a/spec/unit/authorization/authorization_adapter_spec.rb
+++ b/spec/unit/authorization/authorization_adapter_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::AuthorizationAdapter do
 
   let(:adapter) { ActiveAdmin::AuthorizationAdapter.new(double, double) }