--- a/spec/unit/authorization/controller_authorization_spec.rb
+++ b/spec/unit/authorization/controller_authorization_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe "Controller Authorization", type: :controller do
 
   let(:authorization){ controller.send(:active_admin_authorization) }