--- a/spec/unit/auto_link_spec.rb
+++ b/spec/unit/auto_link_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require 'active_admin/view_helpers/active_admin_application_helper'
 require 'active_admin/view_helpers/auto_link_helper'
 require 'active_admin/view_helpers/display_helper'