--- a/spec/unit/comments_spec.rb
+++ b/spec/unit/comments_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe "Comments" do
   let(:application) { ActiveAdmin::Application.new }