--- a/spec/unit/component_spec.rb
+++ b/spec/unit/component_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 class MockComponentClass < ActiveAdmin::Component; end
 
 RSpec.describe ActiveAdmin::Component do