--- a/spec/unit/controller_filters_spec.rb
+++ b/spec/unit/controller_filters_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Application do
   let(:application){ ActiveAdmin::Application.new }
   let(:controllers){ application.controllers_for_filters }