--- a/spec/unit/devise_spec.rb
+++ b/spec/unit/devise_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Devise::Controller do
 
   let(:controller_class) do