--- a/spec/unit/dynamic_settings_spec.rb
+++ b/spec/unit/dynamic_settings_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::DynamicSettingsNode do
   subject { ActiveAdmin::DynamicSettingsNode.build }