--- a/spec/unit/filters/active_spec.rb
+++ b/spec/unit/filters/active_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Filters::Active do
 
   let(:resource) do