--- a/spec/unit/form_builder_spec.rb
+++ b/spec/unit/form_builder_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require "rspec/mocks/standalone"
 
 RSpec.describe ActiveAdmin::FormBuilder do