--- a/spec/unit/i18n_spec.rb
+++ b/spec/unit/i18n_spec.rb
@@ -1,4 +1,5 @@
 require 'i18n/tasks'
+require 'i18n-spec'
 
 Dir.glob('config/locales/*.yml') do |locale_file|
   RSpec.describe locale_file do