--- a/spec/unit/menu_item_spec.rb
+++ b/spec/unit/menu_item_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require 'active_admin/menu'
 require 'active_admin/menu_item'