--- a/spec/unit/page_controller_spec.rb
+++ b/spec/unit/page_controller_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::PageController do
   let(:controller) { ActiveAdmin::PageController.new }
 end