--- a/spec/unit/page_spec.rb
+++ b/spec/unit/page_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require File.expand_path('config_shared_examples', File.dirname(__FILE__))
 
 module ActiveAdmin