--- a/spec/unit/pretty_format_spec.rb
+++ b/spec/unit/pretty_format_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require 'active_admin/view_helpers/display_helper'
 
 RSpec.describe "#pretty_format" do