--- a/spec/unit/pundit_adapter_spec.rb
+++ b/spec/unit/pundit_adapter_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 class DefaultPolicy < ApplicationPolicy
   def respond_to_missing?(method, include_private = false)
     method.to_s[0...3] == "foo" || super