--- a/spec/unit/resource/ordering_spec.rb
+++ b/spec/unit/resource/ordering_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 module ActiveAdmin
   RSpec.describe Resource, "Ordering" do
     describe "#order_by" do