--- a/spec/unit/resource/pagination_spec.rb
+++ b/spec/unit/resource/pagination_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 module ActiveAdmin
   RSpec.describe Resource, "Pagination" do