--- a/spec/unit/resource/scopes_spec.rb
+++ b/spec/unit/resource/scopes_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 module ActiveAdmin
   RSpec.describe Resource, "Scopes" do