--- a/spec/unit/resource/sidebars_spec.rb
+++ b/spec/unit/resource/sidebars_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Resource::Sidebars do
 
   let(:resource) do