--- a/spec/unit/resource_controller/sidebars_spec.rb
+++ b/spec/unit/resource_controller/sidebars_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::ResourceController::Sidebars, type: :controller do
   let(:klass){ Admin::PostsController }