--- a/spec/unit/settings_node_spec.rb
+++ b/spec/unit/settings_node_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::SettingsNode do
   subject { ActiveAdmin::SettingsNode.build }
   let!(:child) { ActiveAdmin::SettingsNode.build(subject) }