--- a/spec/unit/view_factory_spec.rb
+++ b/spec/unit/view_factory_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 def it_should_have_view(key, value)
   it "should have #{value} for view key '#{key}'" do
     expect(subject.send(key)).to eq value