--- a/spec/unit/view_helpers/flash_helper_spec.rb
+++ b/spec/unit/view_helpers/flash_helper_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::ViewHelpers::FlashHelper do
 
   describe '.flash_messages' do