--- a/spec/unit/view_helpers/method_or_proc_helper_spec.rb
+++ b/spec/unit/view_helpers/method_or_proc_helper_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe MethodOrProcHelper do
   let(:receiver) { double }