--- a/spec/unit/views/components/batch_action_selector_spec.rb
+++ b/spec/unit/views/components/batch_action_selector_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require 'active_admin/batch_actions/views/batch_action_selector'
 
 RSpec.describe ActiveAdmin::BatchActions::BatchActionSelector do