--- a/spec/unit/views/components/sidebar_spec.rb
+++ b/spec/unit/views/components/sidebar_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Views::Sidebar do
 
   let(:section) do