--- a/spec/unit/views/index_as_blog_spec.rb
+++ b/spec/unit/views/index_as_blog_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 include ActiveAdmin
 RSpec.describe ActiveAdmin::Views::IndexAsBlog do
   subject { described_class.new }