--- a/spec/unit/views/pages/index_spec.rb
+++ b/spec/unit/views/pages/index_spec.rb
@@ -1,3 +1,5 @@
+require 'rails_helper'
+
 RSpec.describe ActiveAdmin::Views::Pages::Index do
   describe "#title" do
     let!(:application){ ActiveAdmin::Application.new }