This is part of #5514.

The idea here is that not all specs need to load the full Rails environment. When those are run in isolation, the changes here make them run faster. In particular, I think it might not be uncommon for contributors to run the changelog spec in isolation, so I guess it's worth making it faster.

The difference is pretty noticiable:

### Before

```
$ time bundle exec rspec spec/changelog_spec.rb 
Run options:
  include {:focus=>true}
  exclude {:skip=>true}

All examples were filtered out; ignoring {:focus=>true}

Randomized with seed 8885
..

Finished in 0.09076 seconds (files took 5.24 seconds to load)
2 examples, 0 failures

Randomized with seed 8885


real	0m6,242s
user	0m4,424s
sys	0m0,927s
```

### After

```
$ time bundle exec rspec spec/changelog_spec.rb 
Run options:
  include {:focus=>true}
  exclude {:skip=>true}

All examples were filtered out; ignoring {:focus=>true}

Randomized with seed 211
..

Finished in 0.00661 seconds (files took 0.10919 seconds to load)
2 examples, 0 failures

Randomized with seed 211


real	0m0,720s
user	0m0,674s
sys	0m0,048s
```